Updated "Westeros library" after Pasha's review.
